refactor menu method to menu.rb file
