+++ Gemfile
@@ -1,6 +1,8 @@
 source 'https://rubygems.org'
 
+gem 'colorize'
 gem 'launchy'
 gem 'rspec'
 gem 'rubocop', '~>0.81.0'
 gem 'terminal-table'
+gem 'colorize'
